requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
asyncpg==0.21.0
billiard==3.6.3.0
celery==5.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
hiredis==1.1.0
kombu==5.0.2
llvmlite==0.35.0
numba==0.52.0
numpy==1.19.4
pandas==1.1.4
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
zstandard==0.14.0
